requirements.txt
tornado==6.4.2
mysql-connector-python==9.2.0
PyJWT==2.10.1
PyYAML==6.0.2